+++ src/include/GuiListener.h
@@ -9,7 +9,7 @@ class GuiListener
 public:
 	virtual EffectBase* addEffect(EffectIndx desired_effect) = 0;
 	virtual void removeEffect(EffectBase* effect) = 0;
-	virtual void alterEffect(EffectBase* effect, ParamIndx parameter, int32_t new_val) = 0;
+	virtual void alterEffect(EffectBase* effect, ParamIndx parameter, float new_val) = 0;
 
 protected:
 	bool m_alteringEffects;
